PerceptronModel always works better when not masking actions
Masking actions seems to be a good idea, but if an action is always taken in
a state, the others are always masked (given an error of 0). This allows them
to "drift away" as the neural network learns the non-masked action. After some
time, the other actions may receive a value that is completely different from
what Q-Learning has given them. For instance, a bad action can have its value
increasing over time, and it will finally be taken even if it is not optimal.

Masking actions was responsible for the instability of PerceptronModel.